+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+### Fixed
+
+* Allow URLs where username or password contains unescaped '@'. (#2986)
+
 ## 0.25.2 (24th November, 2023)
 
 ### Added